Refactor filters.
