+++ src/js/services/qws/serverMeta.js
@@ -1,5 +1,4 @@
-import { pluralize, stripNonAscii } from "./text";
-import { calcPlayerDisplay } from "./playerDisplay";
+import { pluralize, stripNonAscii } from "./../../util/text";
 
 export const metaByServer = (server) => {
   let clientNames = server.Players.map((p) => p.Name) + server.SpectatorNames;
@@ -60,6 +59,29 @@ const calcSpectatorText = (spectators) => {
   }
 
   return text;
+};
+
+const calcPlayerDisplay = (server, maxRows) => {
+  const showAsTwoColumns = "1on1" === server.Mode || 2 === server.Teams.length;
+
+  const miscRowCount =
+    Number("matchtag" in server.Settings) + Number(showAsTwoColumns);
+  const maxPlayerRows = Math.max(0, maxRows - miscRowCount);
+
+  const playersPerRow = showAsTwoColumns ? 2 : 1;
+  const totalPlayerRows = Math.ceil(server.PlayerSlots.Used / playersPerRow);
+  const visiblePlayerRows = Math.min(maxPlayerRows, totalPlayerRows);
+
+  const visiblePlayers = Math.min(
+    server.PlayerSlots.Used,
+    visiblePlayerRows * playersPerRow
+  );
+  const hiddenPlayers = server.PlayerSlots.Used - visiblePlayers;
+
+  return {
+    visible: visiblePlayers,
+    hidden: hiddenPlayers,
+  };
 };
 
 const gameTimeProgress = (minutesRemaining) => {
