+++ src/js/services/qws/servers.js
@@ -1,11 +1,34 @@
 import { qwsSlice } from "./qws.js";
 import { createEntityAdapter, createSelector } from "@reduxjs/toolkit";
-
-import { metaByServer } from "../../util/serverMeta.js";
-import { compareServers } from "../../util/sort.js";
+import { metaByServer } from "./serverMeta.js";
+import { filterServers } from "./serverFilters.js";
+import { selectUi } from "./../../uiSlice.js";
 
 const serversAdapter = createEntityAdapter({
   selectId: (server) => server.Address,
+  sortComparer: (a, b) => {
+    if (a.meta.score > b.meta.score) {
+      return -1;
+    } else if (a.meta.score < b.meta.score) {
+      return 1;
+    }
+
+    // player slots
+    if (a.PlayerSlots.Free < b.PlayerSlots.Free) {
+      return -1;
+    } else if (a.PlayerSlots.Free > b.PlayerSlots.Free) {
+      return 1;
+    }
+
+    // Address
+    if (a.Address < b.Address) {
+      return -1;
+    } else if (a.Address > b.Address) {
+      return 1;
+    }
+
+    return 0;
+  },
 });
 
 const initialState = serversAdapter.getInitialState();
@@ -35,21 +58,24 @@ export const serversSlice = qwsSlice.injectEndpoints({
           servers[i].meta = metaByServer(servers[i]);
         }
 
-        servers.sort(compareServers);
-
         return serversAdapter.setAll(initialState, servers);
       },
     }),
   }),
 });
 
-export const selectserversResult = serversSlice.endpoints.getMvdsv.select({});
-export const selectserversData = createSelector(
-  selectserversResult,
+export const selectServersResult = serversSlice.endpoints.getMvdsv.select({});
+export const selectServersData = createSelector(
+  [selectServersResult],
   (result) => result.data
 );
 
 export const { selectAll: selectAllServers, selectById: selectServerById } =
   serversAdapter.getSelectors(
-    (state) => selectserversData(state) ?? initialState
+    (state) => selectServersData(state) ?? initialState
   );
+
+export const selectFilteredServers = createSelector(
+  [selectAllServers, selectUi],
+  (servers, ui) => filterServers(servers, ui.filters, ui.favorites)
+);
